Add capabilities for the TTS + ChatGPT to be sent to a GPU server to produce an animation.
This produced an mp4 animation for a face bmp + a mp3 file of the
TTS that ChatGPT is trying to make. This allows you, therefore, to
basically talk to a face. Has some bugs, like the fact that the image size
isn't constant, so that has to change as well as the fact that the
timing (sync of audio and video) is slightly off (and gets pretty bad for
long answers). It also takes ~20 sec to respond to you, but with that said
it is e2e and works.

Parent dies => children die in multiproc

wip - state management seems OK

Might ahve troubles with async/parallel network awaitall... rruh roh

Quick bugfix is_posix

network test works, slightly behind sync with words (video) but whatever

Basic features work